Rename crypto.h to cose_crypto.h
This is needed to make MBed-Studio happy because it does a really bad job of setting the include path
