+++ src/Cose.c
@@ -6,7 +6,7 @@
 #include "cose/cose.h"
 #include "cose/cose_configure.h"
 #include "cose_int.h"
-#include "crypto.h"
+#include "cose_crypto.h"
 
 bool IsValidCOSEHandle(HCOSE h)
 {
