+++ src/CounterSign0.c
@@ -6,7 +6,7 @@
 #include "cose/cose.h"
 #include "cose_int.h"
 #include "cose/cose_configure.h"
-#include "crypto.h"
+#include "cose_crypto.h"
 
 #if USE_COUNTER_SIGNATURE1
