+++ src/Encrypt.c
@@ -15,7 +15,7 @@
 #include "cose/cose.h"
 #include "cose_int.h"
 #include "cose/cose_configure.h"
-#include "crypto.h"
+#include "cose_crypto.h"
 
 #if INCLUDE_ENCRYPT || INCLUDE_MAC
 COSE *EnvelopedRoot = NULL;
