+++ src/Encrypt0.c
@@ -13,7 +13,7 @@
 #include "cose/cose.h"
 #include "cose_int.h"
 #include "cose/cose_configure.h"
-#include "crypto.h"
+#include "cose_crypto.h"
 
 #if INCLUDE_ENCRYPT0 || INCLUDE_MAC0
