+++ src/SignerInfo.c
@@ -11,7 +11,7 @@
 #include "cose/cose.h"
 #include "cose_int.h"
 #include "cose/cose_configure.h"
-#include "crypto.h"
+#include "cose_crypto.h"
 
 #if INCLUDE_SIGN || INCLUDE_COUNTERSIGNATURE
