+++ src/cose_crypto.h
